+++ lib/src/main/java/com/github/hussainderry/securepreferences/model/EncryptionAlgorithm.java
@@ -10,6 +10,9 @@ public enum EncryptionAlgorithm {
 
     AES(new int[]{128, 196, 256}),
 
+    /** ChaCha20-Poly1305 authenticated encryption. 256-bit key only. */
+    CHACHA20(new int[]{256}),
+
     /**
      * @deprecated TripleDES is deprecated by NIST (SP 800-67 Rev 2) and will be
      * removed in v6. Migrate existing data to AES using